+++ cmd/metal-api/internal/service/v1/partition.go
@@ -6,7 +6,7 @@ import (
 
 type PartitionBase struct {
 	MgmtServiceAddress         *string `json:"mgmtserviceaddress" description:"the address to the management service of this partition" optional:"true"`
-	PrivateNetworkPrefixLength *int    `json:"privatenetworkprefixlength" description:"the length of private networks for the machine's child networks in this partition, default 22" optional:"true" minimum:"16" maximum:"30"`
+	PrivateNetworkPrefixLength *uint8  `json:"privatenetworkprefixlength" description:"the length of private networks for the machine's child networks in this partition, default 22" optional:"true" minimum:"16" maximum:"30"`
 }
 
 type PartitionBootConfiguration struct {
